Merges "Added a button to copy a permalink to this tune." (#20)
* Allow arbitrary amounts of whitespace

While editing a tune, it's useful to add newlines and extra spaces. With this commit, the page should now properly handle such redundant whitespace.

Note: this commit has been made directly in the GitHub web interface.

* npm run build

* Added a button to copy a permalink to this tune.

We now have two "copy" buttons: one to copy the tune for pasting into
GRUB (which means removing all excess whitespace), and another to copy a
permalink to the current tune in the current page. This second one is
handful when people share tunes, such as in this gist:
https://gist.github.com/ArtBIT/cfb030c0791b42330381acce33f82ca0

---------

Co-authored-by: Felipe Peñailillo <BreadMaker@users.noreply.github.com>